Change the DB-ID-Name and the len to int of tables. Update bootstrap to 4.3.1. Add the DB Person class.

+++ src/jmbde/Pages/ChipCardDoors/Edit.cshtml.cs
@@ -83,7 +83,7 @@ namespace JMuelbert.BDE.Pages.ChipCardDoors {
         /// </summary>
         /// <returns>The get async.</returns>
         /// <param name="id">Identifier.</param>
-        public async Task<IActionResult> OnGetAsync (long? id) {
+        public async Task<IActionResult> OnGetAsync (int? id) {
             _logger.LogDebug ($"ChipCardDoors/Edit/OnGetAsync({ id })");
 
             if (id == null) {
@@ -104,7 +104,7 @@ namespace JMuelbert.BDE.Pages.ChipCardDoors {
         /// </summary>
         /// <param name="id"></param>
         /// <returns></returns>
-        public async Task<IActionResult> OnPostAsync (long? id) {
+        public async Task<IActionResult> OnPostAsync (int? id) {
             _logger.LogDebug ($"ChipCardDoors/Edit/OnPostAsync({ id })");
 
             if (!ModelState.IsValid) {
