+++ src/jmbde/Pages/ZipCodes/Edit.cshtml.cs
@@ -78,7 +78,7 @@ namespace JMuelbert.BDE.Pages.ZipCodes {
         [BindProperty]
         public ZipCode ZipCode { get; set; }
 
-        public async Task<IActionResult> OnGetAsync (long? id) {
+        public async Task<IActionResult> OnGetAsync (int? id) {
             _logger.LogDebug ($"ZipCodes/Edit/OnGetAsync({ id })");
 
             if (id == null) {
@@ -98,7 +98,7 @@ namespace JMuelbert.BDE.Pages.ZipCodes {
         /// </summary>
         /// <returns>The get async.</returns>
         /// <param name="id">Identifier.</param>
-        public async Task<IActionResult> OnPostAsync (long? id) {
+        public async Task<IActionResult> OnPostAsync (int? id) {
             _logger.LogDebug ($"ZipCodes/Edit/OnPostAsync({ id })");
 
             if (!ModelState.IsValid) {
